refactor: merge api test files
