--- sda/cmd/api/api_integgration_test.go
@@ -1,328 +1,0 @@
-package main
-
-import (
-	"database/sql"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"path"
-	"runtime"
-	"strconv"
-	"testing"
-	"time"
-
-	"github.com/neicnordic/sensitive-data-archive/internal/broker"
-	"github.com/neicnordic/sensitive-data-archive/internal/config"
-	"github.com/neicnordic/sensitive-data-archive/internal/database"
-	"github.com/neicnordic/sensitive-data-archive/internal/helper"
-
-	_ "github.com/lib/pq"
-	log "github.com/sirupsen/logrus"
-
-	"github.com/gin-gonic/gin"
-	"github.com/ory/dockertest/v3"
-	"github.com/ory/dockertest/v3/docker"
-	"github.com/stretchr/testify/assert"
-)
-
-var dbPort, mqPort, OIDCport int
-
-func TestMain(m *testing.M) {
-	if _, err := os.Stat("/.dockerenv"); err == nil {
-		m.Run()
-	}
-
-	_, b, _, _ := runtime.Caller(0)
-	rootDir := path.Join(path.Dir(b), "../../../")
-
-	// uses a sensible default on windows (tcp/http) and linux/osx (socket)
-	pool, err := dockertest.NewPool("")
-	if err != nil {
-		log.Fatalf("Could not construct pool: %s", err)
-	}
-
-	// uses pool to try to connect to Docker
-	err = pool.Client.Ping()
-	if err != nil {
-		log.Fatalf("Could not connect to Docker: %s", err)
-	}
-
-	// pulls an image, creates a container based on it and runs it
-	// TODO use sda-db or postgres repo?
-	postgres, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: "ghcr.io/neicnordic/sda-db",
-		Tag:        "v2.1.3",
-		Env: []string{
-			"DB_LEGA_IN_PASSWORD=lega_in",
-			"DB_LEGA_OUT_PASSWORD=lega_out",
-			"NOTLS=true",
-			"POSTGRES_PASSWORD=rootpassword",
-		},
-	}, func(config *docker.HostConfig) {
-		// set AutoRemove to true so that stopped container goes away by itself
-		config.AutoRemove = true
-		config.RestartPolicy = docker.RestartPolicy{
-			Name: "no",
-		}
-	})
-	if err != nil {
-		log.Fatalf("Could not start resource: %s", err)
-	}
-
-	dbHostAndPort := postgres.GetHostPort("5432/tcp")
-	dbPort, _ = strconv.Atoi(postgres.GetPort("5432/tcp"))
-	// TODO this url or postgres://postgres:rootpasswd@%s/sda?sslmode=disable"
-	databaseURL := fmt.Sprintf("postgres://lega_in:lega_in@%s/lega?sslmode=disable", dbHostAndPort)
-
-	pool.MaxWait = 120 * time.Second
-	if err = pool.Retry(func() error {
-		db, err := sql.Open("postgres", databaseURL)
-		if err != nil {
-			log.Println(err)
-
-			return err
-		}
-
-		return db.Ping()
-	}); err != nil {
-		log.Fatalf("Could not connect to postgres: %s", err)
-	}
-
-	// TODO use sda-mq or rabbitmq?
-	// pulls an image, creates a container based on it and runs it
-	rabbitmq, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: "ghcr.io/neicnordic/sda-mq",
-		Tag:        "v1.4.30",
-		Env: []string{
-			"MQ_USER=test",
-			"MQ_PASSWORD_HASH=C5ufXbYlww6ZBcEqDUB04YdUptO81s+ozI3Ll5GCHTnv8NAm",
-			"MQ_VHOST=test",
-			"NOTLS=true",
-		},
-	}, func(config *docker.HostConfig) {
-		// set AutoRemove to true so that stopped container goes away by itself
-		config.AutoRemove = true
-		config.RestartPolicy = docker.RestartPolicy{
-			Name: "no",
-		}
-	})
-	if err != nil {
-		log.Fatalf("Could not start resource: %s", err)
-	}
-
-	mqPort, _ = strconv.Atoi(rabbitmq.GetPort("5672/tcp"))
-	mqHostAndPort := rabbitmq.GetHostPort("15672/tcp")
-
-	client := http.Client{Timeout: 5 * time.Second}
-	req, err := http.NewRequest(http.MethodGet, "http://"+mqHostAndPort+"/api/users", http.NoBody)
-	if err != nil {
-		log.Fatal(err)
-	}
-	req.SetBasicAuth("test", "test")
-
-	// exponential backoff-retry, because the application in the container might not be ready to accept connections yet
-	if err := pool.Retry(func() error {
-		res, err := client.Do(req)
-		if err != nil {
-			return err
-		}
-		res.Body.Close()
-
-		return nil
-	}); err != nil {
-		if err := pool.Purge(postgres); err != nil {
-			log.Fatalf("Could not purge resource: %s", err)
-		}
-		if err := pool.Purge(rabbitmq); err != nil {
-			log.Fatalf("Could not purge resource: %s", err)
-		}
-		log.Fatalf("Could not connect to rabbitmq: %s", err)
-	}
-
-	RSAPath, _ := os.MkdirTemp("", "RSA")
-	if err := helper.CreateRSAkeys(RSAPath, RSAPath); err != nil {
-		log.Panic("Failed to create RSA keys")
-	}
-	ECPath, _ := os.MkdirTemp("", "EC")
-	if err := helper.CreateECkeys(ECPath, ECPath); err != nil {
-		log.Panic("Failed to create EC keys")
-	}
-
-	// OIDC container
-	oidc, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: "python",
-		Tag:        "3.10-slim",
-		Cmd: []string{
-			"/bin/sh",
-			"-c",
-			"pip install --upgrade pip && pip install aiohttp Authlib joserfc requests && python -u /oidc.py",
-		},
-		ExposedPorts: []string{"8080"},
-		Mounts: []string{
-			fmt.Sprintf("%s/.github/integration/sda/oidc.py:/oidc.py", rootDir),
-		},
-	}, func(config *docker.HostConfig) {
-		// set AutoRemove to true so that stopped container goes away by itself
-		config.AutoRemove = true
-		config.RestartPolicy = docker.RestartPolicy{
-			Name: "no",
-		}
-	})
-	if err != nil {
-		log.Fatalf("Could not start resource: %s", err)
-	}
-
-	OIDCport, _ = strconv.Atoi(oidc.GetPort("8080/tcp"))
-	OIDCHostAndPort := oidc.GetHostPort("8080/tcp")
-
-	client = http.Client{Timeout: 5 * time.Second}
-	req, err = http.NewRequest(http.MethodGet, "http://"+OIDCHostAndPort+"/jwk", http.NoBody)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	// exponential backoff-retry, because the application in the container might not be ready to accept connections yet
-	if err := pool.Retry(func() error {
-		res, err := client.Do(req)
-		if err != nil {
-			return err
-		}
-		res.Body.Close()
-
-		return nil
-	}); err != nil {
-		if err := pool.Purge(oidc); err != nil {
-			log.Panicf("Could not purge oidc resource: %s", err)
-		}
-		log.Panicf("Could not connect to oidc: %s", err)
-	}
-
-	log.Println("starting tests")
-	_ = m.Run()
-
-	log.Println("tests completed")
-	if err := pool.Purge(postgres); err != nil {
-		log.Fatalf("Could not purge resource: %s", err)
-	}
-	if err := pool.Purge(rabbitmq); err != nil {
-		log.Fatalf("Could not purge resource: %s", err)
-	}
-	if err := pool.Purge(oidc); err != nil {
-		log.Fatalf("Could not purge resource: %s", err)
-	}
-}
-
-func TestShutdown(t *testing.T) {
-	Conf = &config.Config{}
-	Conf.Broker = broker.MQConf{
-		Host:       "localhost",
-		Port:       mqPort,
-		User:       "test",
-		Password:   "test",
-		RoutingKey: "test",
-		Exchange:   "sda",
-		Ssl:        false,
-		Vhost:      "/test",
-	}
-	Conf.API.MQ, err = broker.NewMQ(Conf.Broker)
-	assert.NoError(t, err)
-
-	Conf.Database = database.DBConf{
-		Host:     "localhost",
-		Port:     dbPort,
-		User:     "lega_in",
-		Password: "lega_in",
-		Database: "lega",
-		SslMode:  "disable",
-	}
-	Conf.API.DB, err = database.NewSDAdb(Conf.Database)
-	assert.NoError(t, err)
-
-	// make sure all conections are alive
-	assert.Equal(t, false, Conf.API.MQ.Channel.IsClosed())
-	assert.Equal(t, false, Conf.API.MQ.Connection.IsClosed())
-	assert.Equal(t, nil, Conf.API.DB.DB.Ping())
-
-	shutdown()
-	assert.Equal(t, true, Conf.API.MQ.Channel.IsClosed())
-	assert.Equal(t, true, Conf.API.MQ.Connection.IsClosed())
-	assert.Equal(t, "sql: database is closed", Conf.API.DB.DB.Ping().Error())
-}
-
-func TestReadinessResponse(t *testing.T) {
-	gin.SetMode(gin.ReleaseMode)
-	r := gin.Default()
-	r.GET("/ready", readinessResponse)
-	ts := httptest.NewServer(r)
-	defer ts.Close()
-
-	Conf = &config.Config{}
-	Conf.Broker = broker.MQConf{
-		Host:       "localhost",
-		Port:       mqPort,
-		User:       "test",
-		Password:   "test",
-		RoutingKey: "test",
-		Exchange:   "sda",
-		Ssl:        false,
-		Vhost:      "/test",
-	}
-	Conf.API.MQ, err = broker.NewMQ(Conf.Broker)
-	assert.NoError(t, err)
-
-	Conf.Database = database.DBConf{
-		Host:     "localhost",
-		Port:     dbPort,
-		User:     "lega_in",
-		Password: "lega_in",
-		Database: "lega",
-		SslMode:  "disable",
-	}
-	Conf.API.DB, err = database.NewSDAdb(Conf.Database)
-	assert.NoError(t, err)
-
-	res, err := http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	defer res.Body.Close()
-
-	// close the connection to force a reconnection
-	Conf.API.MQ.Connection.Close()
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusServiceUnavailable, res.StatusCode)
-	defer res.Body.Close()
-
-	// reconnect should be fast so now this should pass
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	defer res.Body.Close()
-
-	// close the channel to force a reconneciton
-	Conf.API.MQ.Channel.Close()
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusServiceUnavailable, res.StatusCode)
-	defer res.Body.Close()
-
-	// reconnect should be fast so now this should pass
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	defer res.Body.Close()
-
-	// close DB connection to force a reconnection
-	Conf.API.DB.Close()
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusServiceUnavailable, res.StatusCode)
-	defer res.Body.Close()
-
-	// reconnect should be fast so now this should pass
-	res, err = http.Get(ts.URL + "/ready")
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	defer res.Body.Close()
-}
